Support for forcibly canceling the overlay from the overlay ref. This is useful for some overlays that only serve informational purposes (such as to provide status information) and thus might need to close themselves without any user interaction.
For example think of displaying a status (such as loading, saving, deleting, updating, etc...) in a snackbar. The user will not and should not interact with these types of overlays as they are only to let the user know that something is happening. If an error occurs during one of these statuses you may wish to show a new overlay with details of the error. In order to do so you would have to forcibly cancel the status overlay because you are not allowed to open more than one overlay at a time.