We're reverting this because we really should have put it on it's own branch to begin with.
Revert "Overlays that weren't styled like bootstrap modals weren't receiving click events because the pointer-events on the overlay container was set to none to allow the click to pass through to the scrim. This worked for bootstrap style modals because the "modal-content" class "helpfully" sets pointer-events back to "auto". Our sole example is a bootstrap modal so we never encountered this bug until trying to actually use the library."

This reverts commit 7da58bebaa8edad510c86359742e7c839ba10937.